Add files via upload
In the training data, there were three images of foods A, B, and C respectively. Telling us that A was more similar in taste to B than C. The aim was to train a neural network to be able to make this distinction based on the images. In the end, the model had an accuracy of 63%.